#286 `query_result.py`
